requirements.txt
anthropic>=0.44.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pdfplumber>=0.10.0
scikit-learn>=1.4.0
numpy>=1.26.0
lxml>=5.1.0
python-dotenv>=1.0.0
